### 2018-04-28 #213 @rtaph
submission: outcomerate
`outcomerate` is a lightweight R package that implements the standard outcome rates for surveys, as defined in the [Standard Definitions](https://www.aapor.org/Standards-Ethics/Standard-Definitions-(1).aspx) of the American Association of Public Opinion Research (AAPOR).

```
Package: outcomerate
Version: 0.0.1
Title: AAPOR Survey Outcome Rates
Description: Standardized survey outcome rate function, including the response rate, contact rate, cooperation rate, and refusal rate.
Authors@R: person("Rafael", "Pilliard Hellwig", email = "rafael.taph@gmail.com", role = c("aut", "cre"))
License: CC0
    forcats,
    survey,
    covr
    Rdpack (>= 0.7)
RdMacros:
    Rdpack
```

-   URL for the package (the development repository, not a stylized html page): https://github.com/rtaph/outcomerate

  * data munging: because it allows researchers to compute rates of interest given specific input data
  * reproducibility: because it implements standardized definitions so that the research community can better compare the quality of survey data

Survey analysts are the primary target. The package has applications for any survey that uses standard disposition codes and requires the calculation of [AAPOR outcome rates](https://www.aapor.org/Standards-Ethics/Standard-Definitions-(1).aspx).

Not to my knowledge.

No inquiry made.
