### 2018-04-30 #213 @karthik
### Editor checks:

---

#### Editor comments

👋 @rtaph. Thanks for this submission. I've run `goodpractice::gp()` on your package and everything comes out clean. The package meets all of our recommended guidelines really well so thanks! 🚀
I'll start working on finding two reviewers who are able to put the package through a bunch more testing, user interface checking etc so you can get additional feedback.

---

Reviewers: Currently seeking
Due date: TBD
