### 2018-05-24 #213 @nealrichardson
- [ ] As the reviewer I confirm that there are no conflicts of interest for me to review this work

Estimated hours spent reviewing: **3**

---

This package is very focused on a specific problem, and it solves it well. It wraps and automates what would otherwise be tedious arithmetic and formulas you'd have to keep looking up.

Here are a few notes on ways I think you can improve the code and documentation of the package.

#### Code functionality

* It seems that I can't get "all the rates" unless I provide a value for `e`. I think I should be able to easily get all of the rates that don't depend on `e` just by calling `outcomerate(my_data)`. Instead, I get the error "The parameter e must be provided for RR3, RR4, REF2, CON2". It would be nice if instead you excluded those 4 metrics and calculated the rest.
* Why would I ever want `return_nd = TRUE`? I don't see any documentation or examples that do that. In my experience it's wise not to add extra arguments and features with no application or demand, even if it's trivial to do---they clutter your interface and create a liability for you to maintain them. So I'd advocate removing it. But, if you determine that it has value, consider returning `numden` before even bothering to calculate `rates` (outcomerate.R#L151).
* In asserters.R, `assert_weight`: what if the weight vector contains missing values? (It errors, but not helpfully.)

#### Code style

* `NULL` is generally a better default value for function arguments than `NA`. You're less likely to accidentally provide `NULL`, `NA` is actually type `logical` and that can be confusing, and `is.null()` is a much more natural check than `is.na()`, which will break your `if()` statements if the input is length > 1.
* `outcomerate.factor <- outcomerate.character` works--no need to define a different method and coerce the data `as.character()`.
* Likewise, `outcomerate.table <- outcomerate.numeric` passes the tests too.
* Doing that, then outcomerate.R#L108-L114 simplifies a bit too because you don't have to munge the `table` objects you get: `freq <- table(x)` (perhaps multiplied by `weight`)
* Speaking of multiplying by `weight`, what's the point of a scalar (length 1) weight when you're calculating proportions? Doesn't it difference out? (Empirically, it does, just experimenting with altering a scalar weight in the tests.)
* outcomerate.R#L116 might be more readable as `freq <- xtabs(weight ~ x)`. This would require adding `stats` to `Imports`, though `stats` is a base package so it's "free" from a dependency perspective.

#### Docs

* I recommend the [`spelling` package](https://github.com/ropensci/spelling) to automatically check your spelling. (It will catch for you a few misspellings in the readme and the vignette.) You can have it run every time you run `CHECK`. To do so, [add this to your DESCRIPTION](https://github.com/nealrichardson/httptest/commit/4f1db7180bc235046f7f74f98d9ad9fd7f24b1d1#diff-35ba4a2677442e210c23a00a5601aba3), and then a `tests/spelling.R` file with [this](https://github.com/nealrichardson/httptest/blob/4f1db7180bc235046f7f74f98d9ad9fd7f24b1d1/tests/spelling.R). `CHECK` will then report all of your misspellings. Once you've fixed them, you can do `spelling::update_wordlist()` to whitelist the other words that you use (like "middleearth") that it will flag as misspelled but that are ok. Note that that test file I linked has a special extra check I like: by default, `spelling` errors won't fail `CHECK`, but I have it check to see if it's running on Travis-CI and fail the build there if there are misspellings. I find that to be a good way to get me to notice spelling errors without risking a CRAN check failure.
* roxygen: I recommend using the markdown support added in roxygen2 v6.0 (see note in "DESCRIPTION" below for how to add). That way, your `\item{}`s can just become `* ` lists. Will make the inline docs more readable.
* Also would help to offer some more words to explain what the different rates are beyond just "RR1" etc. https://www.aapor.org/AAPOR_Main/media/publications/Standard-Definitions20169theditionfinal.pdf starting at p. 60 is very readable and helped me. (That might be a better link to use in your docs, too.) Some summary e.g. that the "RR" are "Response rates", which are (explain that in your words), and that the differences among RR1-6 are in whether partial interviews are included and how the unknown types are factored in. Something like that.
* What is `fmat` and where does it come from? That seems to be where all of the real business is. It should probably be documented and more visible.

#### Vignette

* I'm a bit more cautious about adding package dependencies than some, and it's only in Suggests, but `tidyverse` is a lot, and you probably don't need all of that for your vignette.
* Why do you assume `e = 0.8` in this example? Seems that if the goal is to prevent "claims regarding survey quality to become opaque", this shouldn't be opaque either. Some discussion around this, around how one might estimate or reason about `e` and not just make up a number that suits their goals, would be helpful.
* The "small wrapper function" `or()` has an ambiguous sounding name. Perhaps something like `get_rates()` instead?

* `context()` can be more useful as a descriptive string rather than the file name, which any test failure will report anyway
* (test-params.R) It's good to assert the error message in `expect_error()` since helpful error messages are an important part of a user interface and can serve as a source of documentation, too. (cf. http://style.tidyverse.org/error-messages.html). As I mentioned above, I found at least one condition that errored but not in a way that would have helped me to understand what I had done wrong.

#### DESCRIPTION

* "Description" needs to be more than one complete sentence. This is a good place for CRAN human checks to hold up accepting your package submission, so you should try to give them the prose they want. Here's a few of the gotchas I remind myself of each time I submit a new package: https://github.com/nealrichardson/skeletor/blob/master/inst/pkg/DESCRIPTION#L4-L9
* add `Roxygen: list(markdown = TRUE)`, as suggested above, so you can use the markdown syntax in your roxygen docs.
* "Suggests" includes `survey` but I don't see it used anywhere, though perhaps I'm missing it.
