### 2018-05-25 #213 @carlganz
---

The goal of the `outcomerate` package is to simplify the calculation of AAPOR's different definitions for response rates of surveys. The package is lightweight, and accomplishes its primary goal. The vignette does a good job explaining response rates, and how to use the package. The package also makes good use of 3-dimensional matrices. There are a few areas where the package can be improved most of which @nealrichardson has already mentioned. Here is a list in no particular order:

* Definitely use NULL instead of NA for the default parameters.

* Similarly, for `assert_weights` consider giving a warning if any of the weights are NA

* When I was reading source code I initially didn't understand what `fmat` was. Even though the data structure is not exported, I encourage you to include source code that generated `fmat` in the "data-raw" folder as Hadley suggests [here](http://r-pkgs.had.co.nz/data.html#data-extdata). I think it makes reading the source code easier if the code for all the data structures used is available.

* I don't think the `apply` used [here](https://github.com/rtaph/outcomerate/blob/master/R/outcomerate.R#L148) is necessary. `numden[,1]/numden[,2]` gives you what you want but drops names. Up to you if you want to keep `apply` approach.

* Replacing `tidyverse` with the specific packages used in the vignette (i.e. dplyr, tidyr, ggplot2) is a good idea.

* In the documentation for the fake `middleearth` dataset you should include a description of each variable.

* Please change name of `or` function in vignette to something more informative.

* As Neal already said some of the `outcome.*` definitions can be simplified.

Good work!
