### 2018-06-03 2a178ba
Implement recommendations from ROpenSci peer reviewers
