+++ R/outcomerate.R
@@ -1,62 +1,90 @@
 #' AAPOR Survey Outcome Rates
 #'
 #' Provides standardized outcome rates for surveys, primarily as defined by the
-#' American Association for Public Opinion Research (AAPOR). Details can be
+#' [American Association for Public Opinion Research (AAPOR)](http://www.aapor.org/). Details can be
 #' found in the Standard Definitions manual \insertCite{aapor}{outcomerate}.
 #'
+#' Survey and public opinion research often categorizes interview attempts of
+#' of a survey according to a set of outcome codes as follows:
 #'
-#' The outcome rates are defined as follows:
+#' * I  = Complete interview
+#' * P  = Partial interview
+#' * R  = Refusal and break-off
+#' * NC = Non-contact
+#' * O  = Other
+#' * UH = Unknown if household/occupied housing unit
+#' * UO = Unknown, other
+#' * NE = Known ineligible
 #'
-#' AAPOR Response Rates:
+#' These high-level classes are used to calculate outcome rates that
+#' provide some measure of quality over the fieldwork. These outcome rates
+#' are defined here as follows:
 #'
-#' \itemize{ \item{RR1}{ = I / [(I + P) + (R + NC + O) + (UH + UO)]} \item{RR2}{
-#' = (I + P) / [(I + P) + (R + NC + O) + (UH + UO)]} \item{RR3}{ = I / [(I + P)
-#' + (R + NC + O) + e(UH + UO)]} \item{RR4}{ = (I + P) / [(I + P) + (R + NC + O)
-#' + e(UH + UO)]} \item{RR5}{ = I / [(I + P) + (R + NC + O)]} \item{RR6}{ = (I +
-#' P) / [(I + P) + (R + NC + O)]} }
+#' __AAPOR Response Rate__
 #'
-#' AAPOR Cooperation Rates:
+#' The proportion of your intended sample that participate in the survey.
 #'
-#' \itemize{ \item{COOP1}{ = I / [(I + P) + R + O]} \item{COOP2}{ = (I + P) /
-#' [(I + P) + R + O]} \item{COOP3}{ = I / [(I + P) + R]} \item{COOP4}{ = (I + P)
-#' / [(I + P) + R]} }
+#' * RR1 = I / ((I + P) + (R + NC + O) + (UH + UO))
+#' * RR2 = (I + P) / ((I + P) + (R + NC + O) + (UH + UO))
+#' * RR3 = I / ((I + P) + (R + NC + O) + e(UH + UO))
+#' * RR4 = (I + P) / ((I + P) + (R + NC + O) + e(UH + UO))
+#' * RR5 = I / ((I + P) + (R + NC + O))
+#' * RR6 = (I + P) / ((I + P) + (R + NC + O))
 #'
-#' AAPOR Refusal Rates:
+#' __AAPOR Cooperation Rates__
 #'
-#' \itemize{ \item{REF1}{ = R / [(I + P) + (R + NC + O) + (UH + UO)]}
-#' \item{REF2}{ = R / [(I + P) + (R + NC + O) + e(UH + UO)]} \item{REF3}{ = R /
-#' [(I + P) + (R + NC + O)]} }
+#' The proportion of contacted respondents who participate in the survey.
 #'
-#' AAPOR Contact Rates:
+#' * COOP1 = I / ((I + P) + R + O)
+#' * COOP2 = (I + P) / ((I + P) + R + O)
+#' * COOP3 = I / ((I + P) + R)
+#' * COOP4 = (I + P) / ((I + P) + R)
 #'
[…]
