### 2018-06-03 #213 @rtaph
Gentlemen-- thank you all for the feedback and sorry it has taken a while to respond. I am integrating your points and hope to be able to share a revised version soon. Aiming for end of week.

Here are my replies to each point:

> It seems that I can't get "all the rates" unless I provide a value for e. I think I should be able to easily get all of the rates that don't depend on e just by calling outcomerate(my_data). Instead, I get the error "The parameter e must be provided for RR3, RR4, REF2, CON2". It would be nice if instead you excluded those 4 metrics and calculated the rest.

Yes, I see where you are coming from and I agree in the case of the default where `rate = NULL`. I would like to further address this with an eligibility_rate() function to estimate ‘e’ based on ineligible respondents, “NE”, in the x vector. The reason not to integrate it into the other collection of rates is that if you calculate grouped outcome rates (one of the motivations for the package), ‘e’ should not be calculated using a split-apply-combine approach, but globally.

> Why would I ever want return_nd = TRUE? I don't see any documentation or examples that do that. In my experience it's wise not to add extra arguments and features with no application or demand, even if it's trivial to do---they clutter your interface and create a liability for you to maintain them. So I'd advocate removing it. But, if you determine that it has value, consider returning numden before even bothering to calculate rates (outcomerate.R#L151).

I find it useful to calculate grouped outcome rates while fieldwork is happening and when counts are small. For example, computing outcome rates by researcher helps spot early problems with field work. However, when counts are low at the start of a project, the rates are unstable. Having numerators and denominators can be useful in combination with empirical Bayes methods to smooth out the raw estimates. I plan to write a vignette on this but have not had time yet.

> In asserters.R, assert_weight: what if the weight vector contains missing values? (It errors, but not helpfully.)

Good catch. I will add a case for this.

> NULL is generally a better default value for function arguments than NA. You're less likely to accidentally provide NULL, NA is actually type logical and that can be confusing, and is.null() is a much more natural check than is.na(), which will break your if() statements if the input is length > 1.

Great point. I will make the change.

> outcomerate.factor <- outcomerate.character works--no need to define a different method and coerce the data as.character().

Clever. I’ve made the change.

> Likewise, outcomerate.table <- outcomerate.numeric passes the tests too.

One thing I am concerned about with this change is that the table would be coerced to an unnamed numeric vector, and that the order of the input then suddenly matters to the function output. A table object whose first element is the counts for “I” would not give the same results as a table whose first element is, say, counts for “NC”.

> Doing that, then outcomerate.R#L108-L114 simplifies a bit too because you don't have to munge the table objects you get: freq <- table(x) (perhaps multiplied by weight)

So if I understand you, you are suggesting to have the different methods resolve to the table method. Is that correct? I will look into this.

> Speaking of multiplying by weight, what's the point of a scalar (length 1) weight when you're calculating proportions? Doesn't it difference out? (Empirically, it does, just experimenting with altering a scalar weight in the tests.)

Yes, it does, unless you return the numerator and denominator. Otherwise it makes no difference.

> outcomerate.R#L116 might be more readable as freq <- xtabs(weight ~ x). This would require adding stats to Imports, though stats is a base package so it's "free" from a dependency perspective.

Interesting—I am open to the idea. Will see how this works when I try to address the point about using the table method as the ultimate method called.

> I recommend the spelling package....

Great suggestion.

> roxygen: I recommend using the markdown support added in roxygen2 v6.0 (see note in "DESCRIPTION" below for how to add). That way, your \item{}s can just become * lists. Will make the inline docs more readable.

Much better, yes. Thanks for the suggestion.

> Also would help to offer some more words to explain what the different rates are beyond just "RR1" etc.

Good point. Will add more text to briefly explain what each rate represents.

> What is fmat and where does it come from? That seems to be where all of the real business is. It should probably be documented and more visible.

I will write a separate doc for `fmat`. It is a formula matrix that defines all of the equations from the AAPOR definitions in one place (instead of writing them out separately as code).

> I'm a bit more cautious about adding package dependencies than some, and it's only in Suggests, but tidyverse is a lot, and you probably don't need all of that for your vignette.

Agreed. Will simplify it down to only those packages used.

> Why do you assume e = 0.8 in this example? Seems that if the goal is to prevent "claims regarding survey quality to become opaque", this shouldn't be opaque either. Some discussion around this, around how one might estimate or reason about e and not just make up a number that suits their goals, would be helpful.

You make a fair point. I think once I add the function eligibility_rate() and show how to calculate it based on ineligible respondents, this will no longer be an issue. The 0.8 in the example is purely illustrative for the fictional case.

> The "small wrapper function" or() has an ambiguous sounding name. Perhaps something like get_rates() instead?

Yes, sure.

> context() can be more useful as a descriptive string rather than the file name, which any test failure will report anyway

Yep, good point.

> (test-params.R) It's good to assert the error message in expect_error() since helpful error messages are an important part of a user interface and can serve as a source of documentation, too. (cf. http://style.tidyverse.org/error-messages.html). As I mentioned above, I found at least one condition that errored but not in a way that would have helped me to understand what I had done wrong.

Yes, good suggestion.

> "Description" needs to be more than one complete sentence. This is a good place for CRAN human checks to hold up accepting your package submission, so you should try to give them the prose they want.

Agreed. Thanks for the catch.

> add Roxygen: list(markdown = TRUE), as suggested above, so you can use the markdown syntax in your roxygen docs.

Thanks.

> "Suggests" includes survey but I don't see it used anywhere, though perhaps I'm missing it.

I was originally going to use a dataset from the package but then changed my mind and made my own `middleearth` dataset. I will remove it.

> Definitely use NULL instead of NA for the default parameters.

Sure.

> Similarly, for assert_weights consider giving a warning if any of the weights are NA

Sure.

> When I was reading source code I initially didn't understand what fmat was. Even though the data structure is not exported, I encourage you to include source code that generated fmat in the "data-raw" folder as Hadley suggests here. I think it makes reading the source code easier if the code for all the data structures used is available.

Yes, sure. I will add a raw data folder and docs for the `fmat` object.

> I don't think the apply used here is necessary. numden[,1]/numden[,2] gives you what you want but drops names. Up to you if you want to keep apply approach.

I think I had implemented it this way to keep the names. I'll take a look at it again.

> Replacing tidyverse with the specific packages used in the vignette (i.e. dplyr, tidyr, ggplot2) is a good idea.

Yes, will do.

> In the documentation for the fake middleearth dataset you should include a description of each variable.

Good suggestion.

> Please change name of or function in vignette to something more informative.

Yes, will do.

> As Neal already said some of the outcome.* definitions can be simplified.
Sure
