### 2018-06-04 #213 @nealrichardson
> > Likewise, outcomerate.table <- outcomerate.numeric passes the tests too.

> One thing I am concerned about with this change is that the table would be coerced to an unnamed numeric vector, and that the order of the input then suddenly matters to the function output.

As far as I know, this method definition doesn't coerce anything--it just says "use this function". A table behaves like a numeric vector with names, so the code in `outcomerate.numeric` just worked when given a `table`.

When I tried it, all of your tests passed, so if there's some behavior that this changes or alters, it's not anything you're asserting in tests. If it's important, I suggest writing a test and seeing if it fails with this change.
