### 2018-06-07 #213 @rtaph
Dear @nealrichardson, @carlganz, and @karthik,

I have implemented all recommendations to the package (rtaph/outcomerate@2a178bad3f8592f763f1b83075d13dd80ac5d642), save for writing another vignette to explain the use of `return_nd = TRUE` which I hope to do soon. In the mean time, please feel free to review the changes and let me know if you are happy with the current state of the package.

Thanks,
