### 2018-09-01 #213 @rtaph
Thank you @karthik, I will do so shortly. Regarding a blog post, I would be happy to participate.

@nealrichardson and @carlganz, with your permission, I would like to add your names to the DESCRIPTION file as reviewers. Do you have an ORCID you would like me to include?
