### 2018-09-07 89e42d0
Add a section on variance estimation to the introductory vignette
