+++ vignettes/intro-to-outcomerate.Rmd
@@ -176,3 +176,25 @@ middleearth %>%
 ```
 
 In this example, we can see that the contact rate (`CON`) and response rate (`RR`) start to degrade in quality towards day 30. If fieldwork was still continuing, this could be something to look into and attempt to explain and/or redress.
+
+
+## Variance Estimation
+
+To estimate the errors from estimates generated by `outcomerate()`, the simplest approach is to use the normal appromation. Since outcome rates are nothing more than proportions (or nearly so), their standard error is given by $SE(p) = \sqrt{(p(1-p))/n}$.
+
+```{r}
+# first, calculate the outcome rates
+(res <- outcomerate(middleearth$code))
+
+# estimate standard errors using the Normal approximation for proportions 
+se <- sapply(res, function(p) sqrt((p * (1 - p)) / nrow(middleearth)))
+```
+
+With the standard error in hand, we can then construct frequentist confidence intervals:
+
+```{r}
+# calculate 95% confidence intervals
+rbind(res - (se * 1.96), res + (se * 1.96))
+```
+
+Weighted variance estimation in complex surveys require different procedures that go beyond the scope of this vignette. We recommend using `svycontrast()` from the `survey` package to obtain design-based errors that account for elements such as clustering and stratification. Bootstrapping primary sampling units (PSUs) may also be an appropriate method depending on the design at hand. 
