### 2018-09-08 ec1acee
Prepare package for transfer to ropensci
