### 2018-09-10 #213 @stefaniebutland
@rtaph Glad to hear you're interested in contributing a blog post about `outcomerate`. (Thanks for your patience with my delayed response)

This link will give you many examples of blog posts by authors of onboarded packages so you can get an idea of the style and length you prefer: https://ropensci.org/tags/review/.

Here are some technical and editorial guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post. We ask that you submit your draft post via pull request a week before the planned publication date so we can give you some feedback.

Deadline? We can publish 2018-10-02 if you submit your draft by 2018-09-25. Let me know if you prefer a later date.

Happy to answer any questions.
