### 2018-09-29 #213 @rtaph
@karthik, can you help me initiate a `gh-pages` branch to host the `\docs` directory as a pkgdown site?

I am writing the rOpenSci onboarding blog and would like to hyperlink examples in the vignette.
